requirements.txt
# Sapphire Exchange - Unified Architecture Requirements
# Core dependencies for the unified Sapphire Exchange application

# GUI Framework
PyQt5>=5.15.9
qasync>=0.23.0

# Blockchain Integration
PyArweave>=0.6.0
requests>=2.28.0
aiohttp>=3.8.0
solana>=0.30.0
solders>=0.18.0

# Cryptography
cryptography>=3.4.7
ed25519-blake2b>=1.4.1
pynacl>=1.4.0
pycryptodome>=3.20.0
base58>=2.1.1
argon2-cffi>=23.0.0

# BIP39 Wallet Support
mnemonic>=0.20
hdwallet>=2.2.1
bip32>=3.4
ecdsa>=0.18.0

# Data Processing
python-dotenv>=0.19.2
dataclasses-json>=0.5.7
typing-extensions>=4.0.0

# Image Processing (for QR codes)
qrcode>=7.4.2
Pillow>=9.5.0

# HTTP Client
httpx>=0.24.0

# Async Support
asyncio-mqtt>=0.11.1

# Development and Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-mock>=3.10.0
black>=22.0.0
flake8>=5.0.0

# Optional: Performance monitoring
psutil>=5.9.0

# Optional: Logging
structlog>=22.0.0

asyncpg>=0.29.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0